backend/requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dulwich==0.20.42
feedparser==6.0.10
Flask==2.1.2
Flask-Cors==3.0.10
grimoirelab-toolkit==0.2.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
perceval==0.17.17
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.9
waitress==2.1.2
Werkzeug==2.1.2
